Uploading the raw output from the chatbot
For anyone reading this note. This is raw output from a chatbot that I have been working on. Do not take any of it literally. Its intentionally all metaphor and allegory. Don't confuse this with dogma or religion. This is art. It is meant to make you think but not tell you how to think. It is also meant to be a sort of map. I am choosing to stand in the middle of logic and art. Not to judge or label. Just to observe and document.